Add unittest for log_event
